pyproject.toml
[tool.poetry]
name = "laby"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
ascii_graph = "^1.5.1"
numpy = "^1.20"
Pillow = "^8.4.0"
rich = "^10.12.0"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
selenium = "^4.0.0"
jupyter = "^1.0.0"
bokeh = "^2.4.1"
pandas = "^1.3.4"
numba = "^0.54.1"
aiohttp = "^3.8.1"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
